refactor: Extract a Component Book
As Book was a repetitive part of the BookShelf Component, it was extracted and coded to be a Component on its own. This action simplified the main Component BookShelf giving it a more readable nesting and also let us have a Book Component for reuse in case we need it in other parts of the program.