refactor: Update App as selected value changes
Added a function to update the state of the App Component as the user selects a different option in the <select> form of each Book.
The Form takes a user choice and then passes the new value and the book on which the selection is done to its parent Component. From there the values go up the Composition hierarchy till they reach the App Component `setState` that is in charge of re-rendering the UI each time state is changed.
Through `setState` the `books` array in App state is given a new value where the only change is done in the object element that would have a title equal to the one captured by the form.