feat: Add Route and Link Components
Added the BrowserRouter Component to the app to let the correct screen show up listening to changes in the URL.
Added a Link Component in place of the anchor tags to provide declarative paths for navigation throught the app.
Added the Route Component to render the right component on the page based on the current URL path.
Moved the anchor tag in the App Component to the BookShelf Component.