refactor: Match shelves for books in both the search and main page
Added code to make books that appear both in the selection page and in the main page match their shelf.
As the search list is created, each book shelf is settled to 'none'. If any book on the search page also appears in the main page, the book in the selection list is given the same shelf it has on the main page.