refactor: Check for duplicates in search list
Added code to avoid that the same book appeared more than once in the search list.
Unfortunately, because of the way the API is coded, sometimes it can retrieve a book from more sources, changing all its properties identifiers, but displaying the same cover to the user, that has thus the impression of seeing it twice.