feat: Place book in library when user selects a shelf
Added code to move a book, displayed in the search screen, to the library if a user selects a shelf for it.
The program checks if the book is already in the shelf, in which case it does nothing. If it is not there yet and the user selects anything different from `none`, a request is made to the API to update the library. The whole books list is then retrieved again to update the main page with the newly inserted item appearing in its chosen shelf.
The query list is deleted either if the user clicks the link arrow or if he/she clicks the `add` button from the main page.