refactor: Modify search page URL and form width
Changed the URL in the browser's address bar to '/search' when the search page is displayed.
Modified the input form width to take the entire width of its parent div, as before it only took a limited width, getting transparent to the books images as the bar was scrolled down.