refactor: Display matching books are user types
Added code to let the search page display books that match a query as the user types.
Now the user doesn't need to hit the Enter button to start a query to the database, but the query starts as soon as the user starts typing into the input field. As the query string changes, the shown results reflect it.