refactor: Empty query as user leaves search page
Refactored code to change the method to empty the search page when a user goes back to the main page.
Previously a click to the 'Back' link called method `emptyQueryArray()` on the BooksApp Component, that emptied the searchResults array.
After the code refactoring in the previous commit, that array is now emptied in the `searchBook()` method that queries the API. Because of that, the only action performed now by the SearchBook Component as a user clicks the `Back` link is emptying the query string.