fix: Display empty screen as query is deleted
Refactored code to always display an empty search screen if there is no query in the input field.
Previously to these changes, if a user inserted a query string and then deleted it very fast, the API wasn't able to follow the stream of inputs. The `searchedBooks` array was not emptied and a page of books was shown that matched the last remaining character.
The SearchBook Component now calls a function whenever the query string in its input field is empty, making the main App Component reflect this input in its state.
Another fix was made in the `handleChange` method of the SelectList Component, checking that a user clicks a value different from `moveTo` before calling the appropriate function in its props.