refactor: Move a book from the search page to some other shelf
Changed code to make a book move to some other shelf even in the search page.
The previous version of the program only allowed to move a book to another shelf from the main page, while on the search page it was possible to add it to the library if it wasn't already there. Now both actions are allowed: adding to the library a book that is not there yet and changing a shelf a book is on.